// frontend/src/api/generated/schemas/etfSubcategoryItem.ts
// @ts-nocheck
// This file is auto-generated by orval. Do not edit manually.
import type { EtfSubcategoryItemChangePct } from "./etfSubcategoryItemChangePct";

/**
 * 单个 ETF 项目 (用于 Tooltip 列表).
 */
export interface EtfSubcategoryItem {
  /** 当日涨跌幅 % */
  change_pct?: EtfSubcategoryItemChangePct;
  /** ETF 代码 */
  code: string;
  /** ETF 名称 */
  name: string;
}